Move shadowAll to main gradle plugin.
